ubports: restore Android hal service
It's not clear why this was disabled; it's an essential service to get
audio working on the Xiaomi Redmi Note 7 Pro (violet).
If it's causing trouble for some devices, an override should be
introduced in these devices' trees, not in this generic image.

Partially reverts aa7f4d5850b8a73a660a4f2a2515d7229465ae49